refactor: move Client::order and Client::market_data into their domain modules
Closes the last standing violation of docs/rules/style/domain-module-layout.md.
Both builder entry points sat in client/sync.rs and client/async.rs while every
sibling entry point (realtime_bars, tick_by_tick, market_depth) lived in its
domain module.

- Client::order -> orders/{sync,async}.rs
- Client::market_data -> market_data/realtime/{sync,async}.rs

Async market_data had no # Examples while its sync twin had two -- a doc-parity
miss tracked in plans/code-consistency-followups.md. Written on arrival, so
# Examples coverage goes 132/152 -> 134/152.

Test coverage moves with the methods: the client-side
builder_factories_are_constructable test narrows to decoder_context (market_data
is already exercised in market_data/builder tests), and orders gains an
order_entry_point_builds_order test on both sides -- Client::order previously had
no domain-side coverage.
