+++ src/client/sync_tests.rs
@@ -3,7 +3,6 @@ use std::sync::Mutex;
 
 use super::*;
 use crate::common::test_utils::helpers::{error_frame, managed_accounts_frame, next_valid_id_frame};
-use crate::contracts::Contract;
 use crate::messages::{IncomingMessages, OutgoingMessages};
 use crate::server_versions;
 use crate::stubs::MessageBusStub;
@@ -52,12 +51,9 @@ fn check_server_version_branches() {
 }
 
 #[test]
-fn builder_factories_are_constructable() {
+fn decoder_context_is_constructable() {
     let client = stubbed_client();
-    let contract = Contract::stock("AAPL").build();
 
-    let _ = client.order(&contract);
-    let _ = client.market_data(&contract);
     let _ = client.decoder_context();
 }
