+++ src/orders/async_tests.rs
@@ -539,3 +539,17 @@ async fn test_global_cancel() {
     assert_eq!(request_message_count(&message_bus), 1);
     assert_request(&message_bus, 0, &global_cancel_request());
 }
+
+#[tokio::test]
+async fn order_entry_point_builds_order() {
+    let message_bus = Arc::new(MessageBusStub::with_responses(vec![]));
+    let client = Client::stubbed(message_bus, server_versions::SIZE_RULES);
+    let contract = Contract::stock("AAPL").build();
+
+    let order = client.order(&contract).buy(100).limit(50.0).build().expect("build failed");
+
+    assert_eq!(order.action, Action::Buy);
+    assert_eq!(order.total_quantity, 100.0);
+    assert_eq!(order.order_type, "LMT");
+    assert_eq!(order.limit_price, Some(50.0));
+}
